requirements.txt
feedparser
PyYAML
aiohttp
aioredis
aiotg
